Variable renaming and small tweaks in decoder to improve compression.
functionally identical, but all variables are renamed so that
it doesn't use any letter not already used in other names.
(e.g. M is fine since we definitely have to use Math, but q isn't)
this improves compression because the DEFLATE tree for
the decoder doesn't need to store lengths for additional literals.

also did a slight reordering to better exploit deduplication:
- documented that an unused `y` in `y=p[C]+=...` is intentional
- made sure that `(C,i)=>` is always followed by `(`
- exploit a common `=t%M` fragment

all combined, this saves ~10B when compressed with DEFLATE.