Much improved model performance with arrayBufferPool.
it turns out that browsers have relatively slow fill methods,
so much that in some cases that can dominate the run time.
this commit introduces the lazy initialization combined with
the "marking" mechanism to ensure that the same array
can be used over and over without the miscompression.
now optimizer is from 20--30% to 3 times faster than before.

<benchmark>
samples used:
- default README inputs for the online demo (5.7 KB text)
- Edge Not Found (32.7 KB JS)

Firefox 92.0b4:
- README: 90s -> 30s
- Edge Not Found: 240s -> 130s

Chrome 92.0:
- README: 22s -> 16s
- Edge Not Found: 85s -> 70s
</benchmark>