Add CompressOptions.inputEndsWithByte.
if given, the compressed input should end with that byte and
it should not appear elsewhere. the decompressor would then
ignore inputLength (set to -1 to prevent a mistake) and read
until that byte is read and added to the output.

this feature is not used by the packer because an embedded
input length it not a problem, but it might exist a case that
bundling inputLength is not worthy and the input can be
tweaked to satisfy the constraint.