Initial WebAssembly compressor implementation.
the compressor is parametrized with compression parameters
so it has to be generated on the fly; this is done by a crazy
code that generates a correct wasm binary from scratch.
I know, this is crazy, but most wasm generators target
C interoperability that we don't really need, and a remaining
few (e.g. binaryen) are too large to fit in Roadroller.