Replace preferTextOverJs with new dynamicModels (-Zdy).
so, preferTextOverJs was a hack made after it was observed
that some JS inputs are better modelled w/o quotes handling.
due to its hackish nature there was no separate option, so
it was specially handled and might cause some confusions:
for example once the input type is set to text in the online
demo the subsequent optimization will never set it back to JS
even when it might be beneficial.

a new dedicated option solves this issue and also allows for
using quotes modelling for texts whenever appropriate.
henceforth the old preferTextOverJs is now deprecated.