Do not insert a newline between the first & second "lines".
as two lines are no longer guaranteed to be separate statements,
the newline is now somewhat useless. all user-facing tools now
do not insert a newline, with a semi-exception of the online demo
which still shows two lines but is a single line after copy & paste.
also made it clear in the README that two "lines" are not required.